requirements.txt
# Core AI and LangGraph
langgraph>=0.5.0
langchain>=0.3.0
langchain-aws>=0.1.0

# AWS Services
boto3>=1.34.0
botocore>=1.34.0

# Web Framework
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.1.0

# HTTP and API
aiohttp>=3.9.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Data Processing
pydantic>=2.5.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0

# Development Tools (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
